requirements.txt
Package                 Version
----------------------- -----------
absl-py                 1.3.0
anyio                   3.6.1
appnope                 0.1.3
argon2-cffi             21.3.0
argon2-cffi-bindings    21.2.0
asttokens               2.0.8
astunparse              1.6.3
attrs                   22.1.0
backcall                0.2.0
beautifulsoup4          4.11.1
bleach                  5.0.1
cachetools              5.2.0
cairocffi               1.4.0
CairoSVG                2.5.2
certifi                 2022.12.7
cffi                    1.15.1
charset-normalizer      2.1.1
contourpy               1.0.6
cssselect2              0.7.0
cycler                  0.11.0
debugpy                 1.6.3
decorator               5.1.1
defusedxml              0.7.1
entrypoints             0.4
executing               1.1.1
fastjsonschema          2.16.2
flatbuffers             22.12.6
fonttools               4.38.0
gast                    0.4.0
google-auth             2.15.0
google-auth-oauthlib    0.4.6
google-pasta            0.2.0
grpcio                  1.51.1
h5py                    3.7.0
idna                    3.4
importlib-metadata      5.0.0
ipykernel               6.16.0
ipython                 8.5.0
ipython-genutils        0.2.0
ipywidgets              8.0.2
jedi                    0.18.1
Jinja2                  3.1.2
joblib                  1.2.0
jsonschema              4.16.0
jupyter                 1.0.0
jupyter_client          7.4.2
jupyter-console         6.4.4
jupyter-core            4.11.1
jupyter-server          1.21.0
jupyterlab-pygments     0.2.2
jupyterlab-widgets      3.0.3
keras                   2.11.0
kiwisolver              1.4.4
libclang                14.0.6
Markdown                3.4.1
MarkupSafe              2.1.1
matplotlib              3.6.2
matplotlib-inline       0.1.6
mediapipe-silicon       0.8.11
mistune                 2.0.4
mlxtend                 0.21.0
nbclassic               0.4.5
nbclient                0.7.0
nbconvert               7.2.1
nbformat                5.7.0
nest-asyncio            1.5.6
notebook                6.5.1
notebook_shim           0.2.0
numpy                   1.23.4
oauthlib                3.2.2
opencv-contrib-python   4.7.0.68
opencv-python           4.6.0.66
opt-einsum              3.3.0
packaging               21.3
pandas                  1.5.2
pandocfilters           1.5.0
parso                   0.8.3
patsy                   0.5.3
pexpect                 4.8.0
pickleshare             0.7.5
Pillow                  9.3.0
pip                     22.3.1
plotly                  5.11.0
prometheus-client       0.15.0
prompt-toolkit          3.0.31
protobuf                4.21.12
psutil                  5.9.2
ptyprocess              0.7.0
pure-eval               0.2.2
pyasn1                  0.4.8
pyasn1-modules          0.2.8
pycairo                 1.23.0
pycparser               2.21
pydotplus               2.0.2
Pygments                2.13.0
pyparsing               3.0.9
pyrsistent              0.18.1
python-dateutil         2.8.2
pytz                    2022.7
pyzmq                   24.0.1
qtconsole               5.3.2
QtPy                    2.2.1
requests                2.28.1
requests-oauthlib       1.3.1
rsa                     4.9
scikit-learn            1.2.0
scikit-multilearn       0.2.0
scipy                   1.9.3
seaborn                 0.12.1
Send2Trash              1.8.0
setuptools              65.7.0
six                     1.16.0
sniffio                 1.3.0
soupsieve               2.3.2.post1
stack-data              0.5.1
statsmodels             0.13.5
tenacity                8.1.0
tensorboard             2.11.0
tensorboard-data-server 0.6.1
tensorboard-plugin-wit  1.8.1
tensorflow-estimator    2.11.0
tensorflow-macos        2.11.0
tensorflow-metal        0.7.0
termcolor               2.1.1
terminado               0.16.0
threadpoolctl           3.1.0
tinycss2                1.2.1
tornado                 6.2
traitlets               5.5.0
typing_extensions       4.4.0
urllib3                 1.26.13
wcwidth                 0.2.5
webencodings            0.5.1
websocket-client        1.4.1
Werkzeug                2.2.2
wheel                   0.38.4
widgetsnbextension      4.0.3
wrapt                   1.14.1
yellowbrick             1.5
zipp                    3.9.0
